refactor: 서비스 계층 분리 #173
- 컨트롤러와 서비스 레이어간의 역할 분리
- response 추가
